fastapi/Oct25/bookmyshow-replica/services/moviesinfo/pyproject.toml
[project]
name = "moviesinfo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "shaikkhajaibrahim", email = "shaikkhajaibrahim@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.121.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.44",
]

[project.scripts]
moviesinfo = "moviesinfo.main:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"
